requirements.txt
python-pptx==0.6.23
PyMuPDF==1.23.8
Pillow==10.1.0
google-genai
python-dotenv
fastapi
uvicorn
python-multipart
